backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-multipart==0.0.20
pillow==10.4.0
setuptools>=70.0.0

# Optional enrichment (Gemini)
google-generativeai==0.8.3
python-dotenv==1.0.1

# OCR (Surya)
# Surya requires PyTorch (install appropriate CPU/CUDA build separately).
surya-ocr==0.17.0

# Legacy OCR (PaddleOCR) - kept for reference; can be removed later
# PaddleOCR currently requires numpy < 2.0
numpy==1.26.4
paddleocr==2.9.1
# Python 3.13 wheels are available for PaddlePaddle 3.x
paddlepaddle==3.2.2
